split simple-web starter config into separate files
