+++ starters/simple-web/src/animations.ts
@@ -1,0 +1,6 @@
+import { createAnimations } from '@tamagui/animations-css'
+
+export const animations = createAnimations({
+  lazy: 'ease-in 500ms',
+  quick: 'ease-in 100ms',
+})
